% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_js.R
\name{example_js}
\alias{example_js}
\title{Launch the standalone helloWidget JavaScript demo}
\usage{
example_js()
}
\description{
Opens the included HTML page in your browser. This page requires the
\code{helloWidget.js} file from the package.
}
